Change the build process of LLVM into a 2-stage build
A previous commit introduced LLVM/clang support that uses
the host compiler (GCC on typical Linux systems) might not
be the best choice for building LLVM's runtimes.

Instead let's split the build into two stages:
* stage 1 will build clang, lld and helper tools
* stage 2 will use the artefacts from stage 1 and build
  the LLVM cross toolchain

There still seems to be an issue in stage 2:
During CMake configuration, a lot of flag tests fail. E.g.:
  [...]
  -- Performing Test COMPILER_RT_HAS_FPIC_FLAG
  -- Performing Test COMPILER_RT_HAS_FPIC_FLAG - Failed
  [...]

This misdetection of available flags can cause follow-up errors
(e.g. relocation errors when building ASAN).
A solution to this issue is not available in this commit.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>